[FIX] Twincity data cleaned
